+++ src/filters/feed_filters.ts
@@ -63,9 +63,11 @@ export function buildNewFilterSettings(): FeedFilterSettings {
 // Compute language, app, etc. tallies for toots in feed and use the result to initialize filter options
 // Note that this shouldn't need to be called when initializing from storage because the filter options
 // will all have been stored and reloaded along with the feed that birthed those filter options.
-export function initializeFiltersWithSummaryInfo(toots: Toot[], userData: UserData): FeedFilterSettings {
-    const filters: FeedFilterSettings = buildNewFilterSettings();
-
+export function updatePropertyFilterOptions(
+    filters: FeedFilterSettings,
+    toots: Toot[],
+    userData: UserData
+): FeedFilterSettings {
     const tootCounts = Object.values(PropertyName).reduce(
         (counts, propertyName) => {
             // Instantiate missing filter sections  // TODO: maybe this should happen in Storage?
@@ -111,4 +113,4 @@ export function initializeFiltersWithSummaryInfo(toots: Toot[], userData: UserDa
     Storage.setFilters(filters);
     traceLog(`[initializeFiltersWithSummaryInfo()] completed, built filters:`, filters);
     return filters;
-}
+};