+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix bug where changes to the feed would invalidate old FeedFilterSettings in the GUI
 * Add `Config.invalidTrendingTags` param to list tags that are too generic to be used
 * `TheAlgorithm.reset()` also resets concurrency